working with tickets
